pyproject.toml
[project]
name = "riski-extraktor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.38.0",
    "gitpython>=3.1.44",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "pytest>=8.4.1",
    "pyyaml>=6.0.2",
    "langchain-docling>=1.0.0",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.0.1",
    "truststore>=0.8.0",
    "httpx>=0.28.1",
    "bs4>=0.0.2",
    "pydantic[email]>=2.11.7",
    "stamina>=25.1.0",
    "pre-commit>=4.2.0",
    "sqlmodel>=0.0.24",
    "psycopg2>=2.9.10",
]

[dependency-groups]
dev = [
    "asyncpg>=0.30.0",
    "ipykernel>=6.29.5",
    "langchain-openai>=0.3.27",
    "langchain-postgres>=0.0.15",
    "pre-commit>=4.2.0",
    "psycopg[binary,pool]>=3.2.9",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "ruff>=0.11.7",
    "truststore>=0.10.1",
]
